refactor(전체): baseball 폴더 생성 후 전체 파일 이동
